+++ HomeDirectAPI/Controllers/MortgageDashboardController.cs
@@ -30,7 +30,7 @@ namespace HomeDirectAPI.Controllers
 
         // GET api/values/5
         [HttpGet("{Year}")]
-        public MortgageDashboardDetails Get(int Year)
+        public MortgageDashboardDetails Get(long Year)
         {
             return repo.Get(Year);
         }
@@ -43,13 +43,13 @@ namespace HomeDirectAPI.Controllers
 
         // PUT api/values/5
         [HttpPut("{id}")]
-        public void Put(int id, [FromBody]string value)
+        public void Put(long id, [FromBody]string value)
         {
         }
 
         // DELETE api/values/5
         [HttpDelete("{id}")]
-        public void Delete(int id)
+        public void Delete(long id)
         {
         }
     }
