+++ HomeDirectAPI/Controllers/UserTypesController.cs
@@ -28,7 +28,7 @@ namespace HomeDirectAPI.Controllers
 
         // GET api/values/5
         [HttpGet("{CatID}")]
-        public UserTypeResponse Get(int CatID)
+        public UserTypeResponse Get(long CatID)
         {
             return repo.Read(CatID);
         }
@@ -49,7 +49,7 @@ namespace HomeDirectAPI.Controllers
 
         // DELETE api/values/5
         [HttpDelete("{CatID}")]
-        public Response Delete(int CatID)
+        public Response Delete(long CatID)
         {
             return repo.Delete(CatID);
         }
