+++ HomeDirectAPI/Controllers/ViewedsController.cs
@@ -30,7 +30,7 @@ namespace HomeDirectAPI.Controllers
 
         // GET api/values/5
         [HttpGet("{id}")]
-        public ViewedResponse Get(int id)
+        public ViewedResponse Get(long id)
         {
             return repo.Read(id);
         }
@@ -51,7 +51,7 @@ namespace HomeDirectAPI.Controllers
 
         // DELETE api/values/5
         [HttpDelete("{id}")]
-        public Response Delete(int id)
+        public Response Delete(long id)
         {
             return repo.Delete(id);
         }
