+++ HomeDirectAPI/Models/Bank.cs
@@ -8,18 +8,18 @@ namespace HomeDirectAPI.Models
     public class Bank
     {
         [Key]
-        public int BankID { get; set; }
+        public long BankID { get; set; }
         public string BankName { get; set; }
         public string HQAddress { get; set; }
         public string BankCategory { get; set; }
-        public int BankAdminUserID { get; set; }
+        public long BankAdminUserID { get; set; }
         [NotMapped]
         public string BankAdminUser { get; set; }
         public string Location { get; set; }
         public string LocationLat { get; set; }
         public string LocationLong { get; set; }
         public decimal Rate { get; set; }
-        public int Balance { get; set; }
+        public long Balance { get; set; }
     }
 
     public class ListBankResponse : Response
