+++ HomeDirectAPI/Models/LoanBuyer.cs
@@ -8,10 +8,10 @@ namespace HomeDirectAPI.Models
     public class LoanBuyer
     {
         [Key]
-        public int LoanBuyerID { get; set; }
+        public long LoanBuyerID { get; set; }
         public string LoanBuyerName { get; set; }
         public string BuyerCategory { get; set; }
-        public int AdminUserID { get; set; }
+        public long AdminUserID { get; set; }
         public string HQAddress { get; set; }
         public string Location { get; set; }
         public string LocationLat { get; set; }
