+++ HomeDirectAPI/Models/LoanMarketShare.cs
@@ -8,7 +8,7 @@ namespace HomeDirectAPI.Models
     public class LoanMarketShare
     {
         [Key]
-        public int ShareID { get; set; }
+        public long ShareID { get; set; }
         public string TitleHolder { get; set; }
         public decimal Amount { get; set; }
         public decimal Repayments { get; set; }
@@ -16,7 +16,7 @@ namespace HomeDirectAPI.Models
         public decimal Rating { get; set; }
         public string Status { get; set; }
         public string Comment { get; set; }
-        public int ProcessedBy { get; set; }
+        public long ProcessedBy { get; set; }
     }
 
     public class ListLoanMarketShareResponse : Response
